# api/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Value Investors Club API",
    "description": "Read-only API for accessing Value Investors Club data",
    "version": "1.0.0"
  },
  "paths": {
...
    "/ideas/": {
      "get": {
        "summary": "Get Ideas",
        "description": "Get investment ideas with optional filtering and sorting by performance.",
        "operationId": "get_ideas_ideas__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "is_short",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Short"
            }
          },
          {
            "name": "is_contest_winner",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Contest Winner"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "has_performance",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Performance"
            }
          },
          {
            "name": "min_performance",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Performance"
            }
          },
          {
            "name": "max_performance",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Performance"
            }
          },
          {
            "name": "performance_period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Which performance period to filter/sort by",
              "default": "one_year_perf",
              "title": "Performance Period"
            },
            "description": "Which performance period to filter/sort by"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Field to sort by. Can be date or performance",
              "default": "date",
              "title": "Sort By"
            },
            "description": "Field to sort by. Can be date or performance"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort order (asc or desc)",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort order (asc or desc)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IdeaResponse"
                  },
                  "title": "Response Get Ideas Ideas  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ideas/{idea_id}": {
      "get": {
        "summary": "Get Idea Detail",
        "description": "Get complete details for a specific idea including related data.",
        "operationId": "get_idea_detail_ideas__idea_id__get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ideas/{idea_id}/performance": {
      "get": {
        "summary": "Get Idea Performance",
        "description": "Get performance metrics for a specific idea including timeline data.",
        "operationId": "get_idea_performance_ideas__idea_id__performance_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ideas/{idea_id}/description": {
      "get": {
        "summary": "Get Idea Description",
        "description": "Get the full description text for an idea.",
        "operationId": "get_idea_description_ideas__idea_id__description_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DescriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ideas/{idea_id}/catalysts": {
      "get": {
        "summary": "Get Idea Catalysts",
        "description": "Get the catalysts text for an idea.",
        "operationId": "get_idea_catalysts_ideas__idea_id__catalysts_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalystsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/debug/ideas/{idea_id}": {
      "get": {
        "summary": "Debug Idea Detail",
        "description": "Simple debug endpoint to get an idea by ID.",
        "operationId": "debug_idea_detail_debug_ideas__idea_id__get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CatalystsResponse": {
        "properties": {
          "catalysts": {
            "type": "string",
            "title": "Catalysts"
          }
        },
        "type": "object",
        "required": [
          "catalysts"
        ],
        "title": "CatalystsResponse",
        "description": "Catalysts for an investment idea."
      },
      "CommentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "posted_at": {
            "type": "string",
            "title": "Posted At"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author",
          "posted_at",
          "text"
        ],
        "title": "CommentResponse",
        "description": "A read-only forum comment captured by the VIC crawler."
      },
      "CompanyResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "company_name"
        ],
        "title": "CompanyResponse",
        "description": "Company information."
      },
...
      "DescriptionResponse": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "DescriptionResponse",
        "description": "Description of an investment idea."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdeaDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link",
            "default": ""
          },
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "is_short": {
            "type": "boolean",
            "title": "Is Short"
          },
          "is_contest_winner": {
            "type": "boolean",
            "title": "Is Contest Winner"
          },
          "company": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DescriptionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "catalysts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatalystsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "performance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PerformanceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            },
            "type": "array",
            "title": "Comments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "user_id",
          "date",
          "is_short",
          "is_contest_winner"
        ],
        "title": "IdeaDetailResponse",
        "description": "Detailed information about an investment idea, including related data."
      },
      "IdeaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link",
            "default": ""
          },
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "is_short": {
            "type": "boolean",
            "title": "Is Short"
          },
          "is_contest_winner": {
            "type": "boolean",
            "title": "Is Contest Winner"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "user_id",
          "date",
          "is_short",
          "is_contest_winner"
        ],
        "title": "IdeaResponse",
        "description": "Basic information about an investment idea."
      },
...
      "UserResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "user_link": {
            "type": "string",
            "title": "User Link"
          }
        },
        "type": "object",
        "required": [
          "username",
          "user_link"
        ],
        "title": "UserResponse",
        "description": "User information."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}